Remove old headers and libraries.
Headers and stub libraries are now auto generated as part of the
build. Remove the old pieces so no one accidentally uses these.

I haven't completely removed the platforms directory because we are
not yet generating the CRT objects.

We don't generate the static libraries yet either, but those
presumably aren't being used in the platform (I guess we'll find out).

Test: make native
Bug: None
Change-Id: I52ea028c363d6c1c5659234054008557bec3235d